requirements.txt
androguard>=3.3.5
beautifulsoup4>=4.9.3
boto3>=1.15.16
chardet>=3.0.4
cryptography>=3.1
cybox>=2.1.0.21
Cython>=0.29.21
Django>=3.1
django-ratelimit>=3.0.1
dnspython>=2.0.0
dpkt>=1.9.3
elasticsearch>=7.9.1
geoip2>=4.0.2
gevent>=20.6.2
git+https://github.com/andreasvc/pyre2.git  # libre2-dev
git+https://github.com/DissectMalware/XLMMacroDeobfuscator.git
https://github.com/volatilityfoundation/volatility/archive/2.6.1.zip
imagehash>=4.1.0
Jinja2>=2.11.2
libvirt-python>=6.6.0  # libvirt-dev
lief>=0.10.1
M2Crypto>=0.36.0
maec>=4.1.0.16
malwareconfig>=1.0.4
matplotlib>=3.3.1
mixbox>=1.0.3
mmbot>=1.0.10
mwcp>=3.1.0
netstruct>=1.1.2
networkx>=2.5
numpy>=1.19.1
olefile>=0.46
oletools>=0.55.1
passlib>=1.7.4
pbkdf2>=1.3
Pebble>=4.5.3
peepdf>=0.4.2
pefile>=2019.4.18
Pillow>=3.2.0
proxmoxer>=1.1.1
pyattck>=2.1.0
pycares>=3.1.1
pyClamd>=0.4.0
pycrypto>=2.6.1
pydeep>=0.4
pyelftools>=0.26
pygal>=2.4.0
pymisp>=2.4.130
pymongo>=3.11.0
PyOpenSSL>=19.1.0
pype32-py3>=0.1
python-dateutil>=2.8.1
python-magic>=0.4.12
python-pcre>=0.7
python-whois>=0.7.3
pyvim>=3.0.2
pyzipper>=0.3.3
regex>=2020.10.11
requests>=2.24.0
scapy>=2.4.4
SFlock>=0.3.10
simplejson>=3.17.2
six>=1.15.0
SQLAlchemy>=1.3.19
tldextract>=2.2.3
#v8py>=0.9.15
WeasyPrint>=51
XenAPI>=2.14
xmltodict>=0.12.0
yara-python>=4.0.2
